% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_school.R
\name{add_school}
\alias{add_school}
\title{Add school to the database}
\usage{
add_school(
  school,
  class = NULL,
  lat,
  long,
  weight,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{school}{the school's name}

\item{class}{the class}

\item{lat}{the latitude of the school}

\item{long}{the longitude of the school}

\item{weight}{a weight for the school}

\item{username}{}

\item{password}{}
}
\value{
nothing
}
\description{
This function adds schools to the database for use in conjunction with the WOW app at BEEF2021
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
