% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_school.R
\name{get_school}
\alias{get_school}
\title{Add school to the database}
\usage{
get_school(username = NULL, password = NULL)
}
\arguments{
\item{username}{}

\item{password}{}
}
\value{
nothing
}
\description{
This function adds schools to the database for use in conjunction with the WOW app at BEEF2021
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
